fix(AndNot): Fix the logic of AndNot (#24)
There was an issue in the logic of AndNot. The difference was computed
from both side, i.e. a.AndNot(b) used to result in all the elements that are not
common to both a and b. It should rather only remove the elements from a
which are also present in b. This PR fixes it.